% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inlineR.R
\name{inlineCxxPlugin}
\alias{inlineCxxPlugin}
\title{Provide needed compilation flags for google performance tools.}
\usage{
inlineCxxPlugin(...)
}
\arguments{
\item{...}{who knows.}
}
\description{
Provide needed compilation flags for google performance tools.
}

